refactor: relocate IsolateNameType
