--- ref_builder/isolate.py
@@ -1,18 +1,0 @@
-from enum import StrEnum
-
-
-class IsolateNameType(StrEnum):
-    """Possible types for isolate names.
-
-    **Ordered by priority**. Do not reorder attributes.
-
-    Isolate name types were previously called "source types". They are referred to this
-    way in Virtool.
-    """
-
-    ISOLATE = "isolate"
-    STRAIN = "strain"
-    CLONE = "clone"
-    VARIANT = "variant"
-    GENOTYPE = "genotype"
-    SEROTYPE = "serotype"
