+++ ref_builder/models/isolate.py
@@ -1,8 +1,24 @@
 from dataclasses import dataclass
+from enum import StrEnum
 
 from pydantic import UUID4, BaseModel
 
-from ref_builder.isolate import IsolateNameType
+
+class IsolateNameType(StrEnum):
+    """Possible types for isolate names.
+
+    **Ordered by priority**. Do not reorder attributes.
+
+    Isolate name types were previously called "source types". They are referred to this
+    way in Virtool.
+    """
+
+    ISOLATE = "isolate"
+    STRAIN = "strain"
+    CLONE = "clone"
+    VARIANT = "variant"
+    GENOTYPE = "genotype"
+    SEROTYPE = "serotype"
 
 
 @dataclass(frozen=True)
