+++ ref_builder/otu/utils.py
@@ -7,9 +7,8 @@ from uuid import UUID
 import structlog
 
 from ref_builder.errors import PlanConformationError
-from ref_builder.isolate import IsolateNameType
 from ref_builder.models.accession import Accession
-from ref_builder.models.isolate import IsolateName
+from ref_builder.models.isolate import IsolateName, IsolateNameType
 from ref_builder.models.plan import Plan, Segment, SegmentRule
 from ref_builder.ncbi.models import NCBIGenbank
 from ref_builder.plan import (
