+++ tests/services/test_isolate.py
@@ -2,8 +2,7 @@ from uuid import uuid4
 
 from pytest_mock import MockerFixture
 
-from ref_builder.isolate import IsolateNameType
-from ref_builder.models.isolate import IsolateName
+from ref_builder.models.isolate import IsolateName, IsolateNameType
 from ref_builder.otu.builders.isolate import IsolateBuilder
 from ref_builder.repo import Repo
 from ref_builder.services.cls import Services
