+++ tests/test_transaction.py
@@ -1,7 +1,6 @@
 from structlog.testing import capture_logs
 
-from ref_builder.isolate import IsolateNameType
-from ref_builder.models.isolate import IsolateName
+from ref_builder.models.isolate import IsolateName, IsolateNameType
 from ref_builder.models.lineage import Lineage, Taxon, TaxonOtherNames
 from ref_builder.models.molecule import Molecule, MoleculeType, Strandedness, Topology
 from ref_builder.models.plan import Plan, Segment
